bcmgenet: Convert receive filter to NetAdapter 2.1
This commit is partial progress towards supporting NetAdapter 2.1 and the
newer builds of Windows that contain it. This change is not backwards
compatible with the current release build of Windows, 19041. Bcmgenet
previously used the NetAdapter 2.0 Preview API for receive filters that is
now deprecated in newer builds, however the driver will not function in 19041
without a receive filter as indicated by an NDIS error.