bcmgenet: Return zero-length unused RX packets
Return packets and fragments remaining in the receive rings when the
adapter stops as zero-length packets rather than ignored packets. This
change resolves an assertion failure in NetAdapter that is not present
in build 19041. This commit has been briefly tested with build 20226
and may not work with earlier or later builds since NetAdapter is under
active development. Commits from this branch may be applied to the rpi4
branch for release builds when the NetAdapter changes in current Insider
builds appear in a release build such as 21H1.